es.uca.modeling.eol: moved transaction statements to EWL wizard
It didn't make all that much sense to create the transaction from the algorithm
itself. Instead, now we create the transaction from the EWL wizards and test
whether the annotation algorithms were successful or not. If they failed for some
of the nodes in the graph, we rollback the entire transaction... or so it should
be, but "abort" is not working as expected. I can Ctrl-Z all the modifications,
so they _are_ being grouped in a transaction, but using "abort" doesn't seem to
roll back the transaction. Perhaps it's an EMC issue?