Time limit inference: fixed point computation is not required
Since we visit paths from most to least restrictive, existing annotations in a
path will be at least as restrictive as the current path, so we can ignore them
and apply the remaining slack to the rest.

This does not require a fixed-point computation: that was just the first attempt
at solving a problem that was finally resolved by sorting the paths from most
to least restrictive. It's best to remove it, as it was introducing unnecessary
complexity in the algorithm.