Concurrent users: better algorithm for reachable nodes
I previously assumed that using 'Activity.all' it would be enough for
getting all the activities in the graph, but this does not apply for
the service composition models. We really need to find all reachable
nodes from the initial node, but the existing helper algorithm is
suboptimal.

I have implemented a simpler algorithm which should take less time and
use up less memory for it.