time_limits.eol: use isKindOf(Activity) to allow subtypes
Using select(r:Activity|r) appears to check that the exact type of r is
exactly Activity, rather than check that r is an Activity or a subtype
of Activity.

This commits ensures we get the right behaviour for time_limits.eol, just
like we did for the concurrent requests inference algorithm.