Time limit inference: improve path ordering
The outgoing edges from a node which receives d seconds and is assigned
a time limit of t are now compared using the remaining time, t-d, rather
than t.

Additionally, the strictest constraint is created from the strictest
constraint among all the paths in the graph, rather than by adding the
constraint of the current node to the strictest constraint among the
outgoing edges of the source.