marte.tests: fix failures when running from Tycho
One of the failures was due to the fact that the MARTE EPackage wasn't
being started in time, and so it wasn't getting registered in the main
EPackage registry.

The other problem is that the tycho-surefire-plugin doesn't like it
when we use Import-Package to only load org.junit, according to this
bug:

http://jira.codehaus.org/browse/SUREFIRE-825

The workaround is to load org.junit.runner through Import-Package as
well. There's the option of including the whole org.junit with
Require-Bundle, but that would bring unnecessary code into the classpath.